+++ Cargo.toml
@@ -8,7 +8,8 @@ members = [
     "chain-signatures/primitives",
     "integration-tests", 
     "chain-signatures/chain-integration-core", 
-    "chain-signatures/chain-canton"
+    "chain-signatures/chain-canton", 
+    "chain-signatures/chain-solana"
 ]
 exclude = ["infra/scripts/generate_keys"]
 resolver = "2"
@@ -81,6 +82,7 @@ mpc-contract = { path = "./chain-signatures/contract" }
 mpc-crypto = { path = "./chain-signatures/crypto" }
 mpc-chain-integration-core = { path = "./chain-signatures/chain-integration-core" }
 mpc-chain-canton = { path = "./chain-signatures/chain-canton" }
+mpc-chain-solana = { path = "./chain-signatures/chain-solana" }
 mpc-keys = { path = "./chain-signatures/keys" }
 mpc-node = { path = "./chain-signatures/node" }
 mpc-primitives = { path = "./chain-signatures/primitives" }
