+++ chain-signatures/node/src/lib.rs
@@ -8,7 +8,6 @@ pub mod indexer;
 
 pub mod indexer_eth;
 pub mod indexer_hydration;
-pub mod indexer_sol;
 pub mod logs;
 pub mod mesh;
 pub mod metrics;
