+++ integration-tests/Cargo.toml
@@ -70,6 +70,7 @@ mpc-crypto.workspace = true
 mpc-contract.workspace = true
 mpc-chain-integration-core.workspace = true
 mpc-chain-canton.workspace = true
+mpc-chain-solana.workspace = true
 mpc-keys.workspace = true
 mpc-node = { path = "../chain-signatures/node", default-features = false, features = [
     "test-feature",
